move fixture and visualization code to new modules
