+++ tests/seg/test_evaluation.py
@@ -1,30 +1,9 @@
 import itk
 import numpy as np
-from typing import Sequence
-from segmantic.prepro.core import Image3
 
 import pytest
 from segmantic.seg import evaluation
-
-
-def make_image(
-    shape: Sequence[int], spacing: Sequence[float], value: int = 0
-) -> Image3:
-    """Create image with specified shape and spacing"""
-    assert len(shape) == len(spacing)
-    dim = len(shape)
-
-    region = itk.ImageRegion[dim]()
-    region.SetSize(shape)
-    region.SetIndex(tuple([0] * dim))
-
-    image = itk.Image[itk.UC, dim].New()
-    image.SetRegions(region)
-    image.SetSpacing(spacing)
-    image.Allocate()
-
-    image[:] = value
-    return image
+from tests.fixture import make_image
 
 
 def test_confusion_matrix():
